A pure component/container design pattern
