+++ templates/container-component.ejs
@@ -1,0 +1,20 @@
+import React, { Component } from 'react'
+import { ScrollView, Text } from 'react-native'
+
+// Styles
+import styles from './Styles/<%= props.name %>Style'
+
+export default class <%= props.name %> extends Component {
+  // constructor (props) {
+  //   super(props)
+  //   this.state = {}
+  // }
+
+  render () {
+    return (
+      <ScrollView style={styles.container}>
+        <Text><%= props.name %> Container</Text>
+      </ScrollView>
+    )
+  }
+}
