+++ templates/container.ejs
@@ -1,26 +1,5 @@
-import React, { Component } from 'react'
-import { ScrollView, Text } from 'react-native'
 import { connect } from 'react-redux'
-// Add Actions - replace 'Your' with whatever your reducer is called :)
-// import YourActions from '../Redux/YourRedux'
-
-// Styles
-import styles from './Styles/<%= props.name %>Style'
-
-class <%= props.name %> extends Component {
-  // constructor (props) {
-  //   super(props)
-  //   this.state = {}
-  // }
-
-  render () {
-    return (
-      <ScrollView style={styles.container}>
-        <Text><%= props.name %> Container</Text>
-      </ScrollView>
-    )
-  }
-}
+import <%= props.name %> from '../Components/<%= props.name %>'
 
 const mapStateToProps = (state) => {
   return {
