+++ templates/flatlist-component.ejs
@@ -1,0 +1,103 @@
+import React from 'react'
+import { View, Text, FlatList } from 'react-native'
+
+// More info here: https://facebook.github.io/react-native/docs/flatlist.html
+
+// Styles
+import styles from './Styles/<%= props.name %>Style'
+
+export default class <%= props.name %> extends React.PureComponent {
+  /* ***********************************************************
+  * STEP 1
+  * This is an array of objects with the properties you desire
+  * Usually this should come from Redux mapStateToProps
+  *************************************************************/
+  state = {
+    dataObjects: [
+      {title: 'First Title', description: 'First Description'},
+      {title: 'Second Title', description: 'Second Description'},
+      {title: 'Third Title', description: 'Third Description'},
+      {title: 'Fourth Title', description: 'Fourth Description'},
+      {title: 'Fifth Title', description: 'Fifth Description'},
+      {title: 'Sixth Title', description: 'Sixth Description'},
+      {title: 'Seventh Title', description: 'Seventh Description'}
+    ]
+  }
+
+  /* ***********************************************************
+  * STEP 2
+  * `renderRow` function. How each cell/row should be rendered
+  * It's our best practice to place a single component here:
+  *
+  * e.g.
+    return <MyCustomCell title={item.title} description={item.description} />
+  *************************************************************/
+  renderRow ({item}) {
+    return (
+      <View style={styles.row}>
+        <Text style={styles.boldLabel}>{item.title}</Text>
+        <Text style={styles.label}>{item.description}</Text>
+      </View>
+    )
+  }
+
+  /* ***********************************************************
+  * STEP 3
+  * Consider the configurations we've set below.  Customize them
+  * to your liking!  Each with some friendly advice.
+  *************************************************************/
+  // Render a header?
+  renderHeader = () =>
+    <Text style={[styles.label, styles.sectionHeader]}> - Header - </Text>
+
+  // Render a footer?
+  renderFooter = () =>
+    <Text style={[styles.label, styles.sectionHeader]}> - Footer - </Text>
+
+  // Show this when data is empty
+  renderEmpty = () =>
+    <Text style={styles.label}> - Nothing to See Here - </Text>
+
+  renderSeparator = () =>
+    <Text style={styles.label}> - ~~~~~ - </Text>
+
+  // The default function if no Key is provided is index
+  // an identifiable key is important if you plan on
+  // item reordering.  Otherwise index is fine
+  keyExtractor = (item, index) => `${index}`
+
+  // How many items should be kept im memory as we scroll?
+  oneScreensWorth = 20
+
+  // extraData is for anything that is not indicated in data
+  // for instance, if you kept "favorites" in `this.state.favs`
+  // pass that in, so changes in favorites will cause a re-render
+  // and your renderItem will have access to change depending on state
+  // e.g. `extraData`={this.state.favs}
+
+  // Optimize your list if the height of each item can be calculated
+  // by supplying a constant height, there is no need to measure each
+  // item after it renders.  This can save significant time for lists
+  // of a size 100+
+  // e.g. itemLayout={(data, index) => (
+  //   {length: ITEM_HEIGHT, offset: ITEM_HEIGHT * index, index}
+  // )}
+
+  render () {
+    return (
+      <View style={styles.container}>
+        <FlatList
+          contentContainerStyle={styles.listContent}
+          data={this.state.dataObjects}
+          renderItem={this.renderRow}
+          keyExtractor={this.keyExtractor}
+          initialNumToRender={this.oneScreensWorth}
+          ListHeaderComponent={this.renderHeader}
+          ListFooterComponent={this.renderFooter}
+          ListEmptyComponent={this.renderEmpty}
+          ItemSeparatorComponent={this.renderSeparator}
+        />
+      </View>
+    )
+  }
+}
